Some compilation fixes for Godot 3.4 (#1979)
* Fix spine-godot README folder paths

* Fix spine-godot compilation issues for godot-3.4

+++ spine-godot/spine_godot/SpineTimeline.cpp
@@ -92,7 +92,7 @@ Array SpineTimeline::getPropertyIds() {
 	res.resize(ids.size());
 
 	for (size_t i = 0; i < res.size(); ++i) {
-		res[i] = ids[i];
+		res[i] = (int64_t) ids[i];
 	}
 
 	return res;
